Fixed management of multiple holidays in the same date
When multiple holidays were defined in the same date, previous versions thrown an error because multiple row with the same date were generated. Now only one holiday is kept and the SubstituteHoliday parameter works correctly in this condition. The new ConflictPriority parameter in HolidayParameters speficy which holiday name to keep in the date marked as holiday in case of multiple holiday names.