Added 4-week periods to FW calendar
The FiscalWeeks calendar (FW prefix) has now a FW Period column that enumerates 13 periods in one year, 4 weeks each one. In years with 53 weeks there is a period P14 with 1 week only.
The FW Periods hierarchy has Year/Period/Week levels - the Week is the End of Week date, so there is a visible column with a date that can be used as axis in a Line Chart.